requirements.txt
certifi==2018.11.29
chardet==3.0.4
dj-database-url==0.5.0
Django==2.1.4
django-multiselectfield==0.1.8
gunicorn==19.9.0
idna==2.8
numpy==1.15.4
pandas==0.23.4
psycopg2-binary==2.7.6.1
python-dateutil==2.7.5
pytz==2018.7
requests==2.21.0
six==1.12.0
urllib3==1.24.1
whitenoise==4.1.2
xlrd==1.1.0
